+++ queries/linked_account/remove.sql
@@ -2,4 +2,4 @@ DELETE FROM
     linked_account
 WHERE
     owner_id = $1
-    AND id = $2;
+    AND id = $2 RETURNING id;