+++ queries/user/delete.sql
@@ -1,4 +1,4 @@
 DELETE FROM
     user_account
 WHERE
-    id = $1;
+    id = $1 RETURNING id;