+++ src/main.rs
@@ -129,6 +129,9 @@ pub struct Config {
     /// S3 bucket.
     #[clap(long, env("S3_BUCKET"), default_value = "fuzzysearch-owo")]
     pub s3_bucket: String,
+    /// S3 CDN prefix, all uploaded files will be $S3_CDN_PREFIX/object/name.
+    #[clap(long, env("S3_CDN_PREFIX"))]
+    pub s3_cdn_prefix: String,
 
     /// Redis DSN, in the format `redis://host/`.
     #[clap(long, env("REDIS_DSN"))]